refactor the country services to a separate class
