+++ src/main/kotlin/io/web/covid19tracker/config/AppConfig.kt
@@ -6,3 +6,7 @@ import org.springframework.context.annotation.Configuration
 @ConfigurationProperties(prefix = "covid19")
 @Configuration
 data class AppConfig(var baseUrl: String = "", var dataUri: String = "")
+
+@ConfigurationProperties(prefix = "covid19api")
+@Configuration
+data class ApiAppConfig(var baseUrl: String = "", var countries: String = "", var totalAllCountries: String = "")
