+++ src/main/kotlin/io/web/covid19tracker/service/DataService.kt
@@ -1,14 +1,7 @@
 package io.web.covid19tracker.service
 
-import com.fasterxml.jackson.core.type.TypeReference
-import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
-import io.web.covid19tracker.config.ApiAppConfig
 import io.web.covid19tracker.config.AppConfig
-import io.web.covid19tracker.models.Country
 import io.web.covid19tracker.models.Data
-import okhttp3.OkHttpClient
-import okhttp3.Request
-import okhttp3.ResponseBody
 import org.apache.commons.csv.CSVFormat
 import org.apache.commons.csv.CSVRecord
 import org.springframework.beans.factory.annotation.Autowired
@@ -21,19 +14,10 @@ import java.io.StringReader
 import javax.annotation.PostConstruct
 
 @Service
-class DataService(
-        @Autowired appConfig: AppConfig,
-        @Autowired apiAppConfig: ApiAppConfig
-) {
+class DataService(@Autowired appConfig: AppConfig) {
     private val baseUrl = appConfig.baseUrl
     private val dataUri = appConfig.dataUri
     private val webClient = WebClient.builder().baseUrl(baseUrl).build()
-
-    private val apiBaseUrl = apiAppConfig.baseUrl
-    private val countriesUri = apiAppConfig.countries
-
-    private val client = OkHttpClient()
-    private val jacksonObjectMapper = jacksonObjectMapper()
 
     @PostConstruct
     @Scheduled(cron = "* * * * * *")
@@ -64,31 +48,6 @@ class DataService(
             val currentCount = it.get(it.size() - 1).toIntOrNull()
             val previousCount = it.get(it.size() - 2).toIntOrNull()
             Data(province, country, currentCount, previousCount)
-        }
-    }
-
-    fun getCountries(): List<Country>? {
-        val countryJson = fetchCountries()?.string()
-        return jacksonObjectMapper.readValue(countryJson, object : TypeReference<List<Country>>() {})
-    }
-
-    private fun fetchCountries(): ResponseBody? {
-        val request = Request
-                .Builder()
-                .url(apiBaseUrl + countriesUri)
-                .get()
-                .build()
-        val httpClient = client.newBuilder().build()
-        val response = httpClient.newCall(request).execute()
-        return response.body
-    }
-
-    fun getCountry(
-            countries: List<Country>?,
-            countryName: String?
-    ): Country? {
-        return countries?.first {
-            it.country == countryName
         }
     }
 }
